requirements.txt
Click==7.0
dj-database-url==0.5.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==0.25.3
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
virtualenv==16.7.9
Werkzeug==0.16.0
